+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/SharedInformerFactory.java
@@ -79,10 +79,10 @@ public class SharedInformerFactory extends BaseOperation {
    * @param apiListTypeClass api list type class
    * @param resyncPeriodInMillis resync period in milliseconds
    * @param <T> the type parameter
-   * @param <TList> the type's list parameter
+   * @param <L> the type's list parameter
    * @return the shared index informer
    */
-  public synchronized <T extends HasMetadata, TList extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<TList> apiListTypeClass, long resyncPeriodInMillis) {
+  public synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<L> apiListTypeClass, long resyncPeriodInMillis) {
     return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, context.withPlural(getPluralFromKind(apiTypeClass.getSimpleName())), resyncPeriodInMillis);
   }
 
@@ -94,13 +94,32 @@ public class SharedInformerFactory extends BaseOperation {
    * @param apiListTypeClass api list type class
    * @param resyncPeriodInMillis resync period in milliseconds
    * @param <T> the type parameter
-   * @param <TList> the type's list parameter
+   * @param <L> the type's list parameter
    * @return the shared index informer
    */
-  public synchronized <T extends HasMetadata, TList extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerForCustomResource(CustomResourceDefinitionContext customResourceContext, Class<T> apiTypeClass, Class<TList> apiListTypeClass, long resyncPeriodInMillis) {
+  public synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerForCustomResource(CustomResourceDefinitionContext customResourceContext, Class<T> apiTypeClass, Class<L> apiListTypeClass, long resyncPeriodInMillis) {
     return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, context.withApiGroupVersion(customResourceContext.getVersion())
       .withApiGroupName(customResourceContext.getGroup())
       .withPlural(customResourceContext.getPlural()), resyncPeriodInMillis);
+  }
+
+  /**
+   * Constructs and returns a shared index informer with resync period specified for custom resources.
+   *
+   * @param customResourceContext basic information about the Custom Resource Definition corresponding to that custom resource
+   * @param apiTypeClass apiType class
+   * @param apiListTypeClass api list type class
+   * @param operationContext operation context
+   * @param resyncPeriodInMillis resync period in milliseconds
+   * @param <T> the type parameter
+   * @param <L> the type's list parameter
+   * @return the shared index informer
+   */
+  public synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerForCustomResource(CustomResourceDefinitionContext customResourceContext, Class<T> apiTypeClass, Class<L> apiListTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
+    return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, this.context.withApiGroupVersion(customResourceContext.getVersion())
+      .withApiGroupName(customResourceContext.getGroup())
+      .withPlural(customResourceContext.getPlural())
+      .withOperationContext(operationContext), resyncPeriodInMillis);
   }
 
   /**
@@ -112,22 +131,22 @@ public class SharedInformerFactory extends BaseOperation {
    * @param operationContext operation context
    * @param resyncPeriodInMillis resync period in milliseconds
    * @param <T> the type parameter
-   * @param <TList> the type's list parameter
+   * @param <L> the type's list parameter
    * @return the shared index informer
    */
-  public synchronized <T extends HasMetadata, TList extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<TList> apiListTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
-    ListerWatcher<T, TList> listerWatcher = listerWatcherFor(apiTypeClass, apiListTypeClass);
-    SharedIndexInformer<T> informer = new DefaultSharedIndexInformer<T, TList>(apiTypeClass, listerWatcher, resyncPeriodInMillis, operationContext, eventListeners);
+  public synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<L> apiListTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
+    ListerWatcher<T, L> listerWatcher = listerWatcherFor(apiTypeClass, apiListTypeClass);
+    SharedIndexInformer<T> informer = new DefaultSharedIndexInformer<>(apiTypeClass, listerWatcher, resyncPeriodInMillis, this.context.withPlural(getPluralFromKind(apiTypeClass.getSimpleName())).withOperationContext(operationContext), eventListeners);
     this.informers.put(apiTypeClass, informer);
     return informer;
   }
 
-  private <T extends HasMetadata, TList extends KubernetesResourceList<T>> ListerWatcher<T, TList> listerWatcherFor(Class<T> apiTypeClass, Class<TList> apiListTypeClass) {
+  private <T extends HasMetadata, L extends KubernetesResourceList<T>> ListerWatcher<T, L> listerWatcherFor(Class<T> apiTypeClass, Class<L> apiListTypeClass) {
 
-    return new ListerWatcher<T, TList>() {
+    return new ListerWatcher<T, L>() {
       @Override
-      public TList list(ListOptions params, String namespace, OperationContext context) throws KubernetesClientException {
-        BaseOperation<T, TList, ?, ?> listBaseOperation = baseOperation.newInstance(context.withNamespace(namespace));
+      public L list(ListOptions params, String namespace, OperationContext context) throws KubernetesClientException {
+        BaseOperation<T, L, ?, ?> listBaseOperation = baseOperation.newInstance(context.withNamespace(namespace));
         listBaseOperation.setType(apiTypeClass);
         listBaseOperation.setListType(apiListTypeClass);
 
@@ -136,7 +155,7 @@ public class SharedInformerFactory extends BaseOperation {
 
       @Override
       public Watch watch(ListOptions params, String namespace, OperationContext context, Watcher<T> resourceWatcher) throws KubernetesClientException {
-        BaseOperation<T, TList, ?, ?> watchBaseOperation = baseOperation.newInstance(context);
+        BaseOperation<T, L, ?, ?> watchBaseOperation = baseOperation.newInstance(context);
         watchBaseOperation.setType(apiTypeClass);
         watchBaseOperation.setListType(apiListTypeClass);